FIX: Kill wait-for-postgres if host or port not set
The script runs forever and simply prints 'waiting for db' if KONG_PG_HOST or
KONG_PG_PORT are not set. Which isn't very helpful.

This commit hard fails if either of those environment variables are not set,
making it easier and quicker to diagnose problems.

As a note, the reason why they may not be set is when a user provides there own
database but doesn't set pg_port or pg_host.

Additional changes:
 - Message is slightly more informative, so if connecting takes awhile,
the use can see host and port being attempted.
 - Set the default postgres port