[PATCH] Ensure environment variables map is ordered (#40)
Since we are dealing with maps, there is zero guarantee that the output yaml will
be the same between runs. In most cases this is not an issue, but the env map is
used by jobs. In this case `spec.template` is immutable and if the env list
order changes, it results in a failed upgrade.

So in the construction of the map, we ensure (by ordering the list of keys) that
it is idempotent.